+++ apps/client/src/pages/search/page/search-page.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useSearchParams } from 'react-router-dom';
 
 import { SearchBar, SearchSuggestionList } from '@confeti/design-system';
-import { SwitchCase } from '@shared/components/SwitchCase';
+import { SwitchCase } from '@shared/components/switch-case';
 import { useDebouncedKeyword } from '@shared/hooks/use-debounce-keyword';
 import { useRelatedSearch } from '@shared/hooks/use-related-search';
 import Loading from '@shared/pages/loading/loading';
