+++ apps/client/src/shared/components/switch-case.tsx
